Phase 7B Session 5: Complete warning elimination in Src/armyG.c
Prompt: Please try Src/armyG.c

Work Completed:
- Fixed 4 warnings in armyG.c (4 → 0 warnings)
- Warning 1: Line 1528 - Cast strlen() result to int for signed/unsigned comparison with COLS expression
- Warning 2: Line 2066 - Changed %d to %ld format specifier for MAX_IDTYPE (long expression)
- Warning 3: Line 2097 - Changed %d to %ld format specifier for MAX_IDTYPE (long expression)
- Warning 4: Line 2129 - Changed %d to %ld format specifier for MAX_IDTYPE (long expression)
- All fixes ensure proper type compatibility while preserving original functionality
- MAX_IDTYPE is defined as ((1L<<(sizeof(idtype) * NBBY)) - 1L) requiring long format specifier
- Clean compilation achieved with zero warnings

+++ Src/armyG.c
@@ -1525,7 +1525,7 @@ ext_armycmd PARM_1 (int, armie)
 	aunit_stat(i)) {
 
       /* check for overruns */
-      if (x + strlen(stat_info[i].input) > COLS - 8) {
+      if (x + (int)strlen(stat_info[i].input) > COLS - 8) {
 	addstr(",");
 	x = 1;
 	move(y++, x);
@@ -2063,7 +2063,7 @@ aunit_numbering PARM_0(void)
       /* check the range */
       if ((x <= 0) ||
 	  (x > MAX_IDTYPE)) {
-	sprintf(string, "A unit number must be between 1 and %d",
+	sprintf(string, "A unit number must be between 1 and %ld",
 		MAX_IDTYPE);
 	errormsg(string);
       } else {
@@ -2094,7 +2094,7 @@ aunit_numbering PARM_0(void)
 	/* check the range */
 	if ((x <= 0) ||
 	    (x > MAX_IDTYPE)) {
-	  sprintf(string, "A unit number must be between 1 and %d",
+	  sprintf(string, "A unit number must be between 1 and %ld",
 		  MAX_IDTYPE);
 	  errormsg(string);
 	} else {
@@ -2126,7 +2126,7 @@ aunit_numbering PARM_0(void)
 	/* check the range */
 	if ((x <= 0) ||
 	    (x > MAX_IDTYPE)) {
-	  sprintf(string, "A unit number must be between 1 and %d",
+	  sprintf(string, "A unit number must be between 1 and %ld",
 		  MAX_IDTYPE);
 	  errormsg(string);
 	} else {
